添加用户认证和授权功能，更新路由和数据库结构

+++ src/routes/user.ts
@@ -2,35 +2,35 @@ import { Hono } from "hono";
 import { db } from "../db/index.js";
 import { users } from "../db/schema.js";
 import { eq } from "drizzle-orm";
-import { z } from "zod";
-import { zValidator } from "@hono/zod-validator";
 
-const createUserSchema = z.object({
-  name: z.string().min(2).max(20),
-  password: z.string().min(6),
-});
-const userRoutes = new Hono();
+import type { AppEnv } from "../types.js";
+import bcrypt from "bcryptjs";
+
+const userRoutes = new Hono<AppEnv>();
 userRoutes.get("/", (c) => {
-  const allUsers = db.select().from(users).all();
-  return c.json(allUsers);
-});
-userRoutes.post("/", zValidator("json", createUserSchema), async (c) => {
-  const body = c.req.valid("json");
-  const exist = db.select().from(users).where(eq(users.name, body.name)).get();
-  if (exist) {
+  const payload = c.get("jwtPayload");
+  console.log(payload);
+  if (payload.name !== "admin") {
     return c.json(
       {
-        error: "用户名已存在",
+        error: "没有权限",
       },
-      400,
+      403,
     );
   }
-  const res = db.insert(users).values(body).returning().get();
-  const { password, ...safeUser } = res;
-  return c.json(safeUser, 201);
+  const allUsers = db.select().from(users).all();
+
+  return c.json(allUsers);
 });
 userRoutes.delete("/:id", (c) => {
   const id = Number(c.req.param("id"));
+  const payload = c.get('jwtPayload')
+
+  if(id !==payload.id){
+    return c.json({
+      error:'没有权限操作'
+    },403)
+  }
   const deleted = db.delete(users).where(eq(users.id, id)).returning().get();
   if (!deleted) {
     return c.json({ error: "用户不存在" }, 404);
@@ -40,6 +40,11 @@ userRoutes.delete("/:id", (c) => {
 });
 userRoutes.patch("/:id", async (c) => {
   const id = Number(c.req.param("id"));
+  const payload = c.get("jwtPayload");
+
+  if (id !== payload.id) {
+    return c.json({ error: "没有权限操作" }, 403);
+  }
   let exist = db.select().from(users).where(eq(users.id, id)).get();
   if (!exist) {
     return c.json(
@@ -50,11 +55,12 @@ userRoutes.patch("/:id", async (c) => {
     );
   }
   const body = await c.req.json();
+  const hashed  =body.password? bcrypt.hashSync(body.password,10):undefined
   const updated = db
     .update(users)
     .set({
       name: body.name ?? exist.name, // 只更新传了的字段
-      password: body.password ?? exist.password,
+      password: hashed ?? exist.password,
     })
     .where(eq(users.id, id))
     .returning()
